[220815-3] Merge branch 'develop' into 'front'

See merge request s07-webmobile2-sub2/S07P12B102!607

+++ backend/src/main/java/com/jourgeois/backend/controller/CocktailController.java
@@ -114,13 +114,17 @@ public class CocktailController {
         }
     }
 
-    @PostMapping(value = "/comment")
-    public ResponseEntity insertReview(@RequestBody CocktailCommentDTO cocktailCommentDTO) {
+    @PostMapping(value = "/auth/comment")
+    public ResponseEntity insertReview(HttpServletRequest request, @RequestBody CocktailCommentDTO cocktailCommentDTO) {
         // uId, cocktailId, review  정보 받음
-        System.out.println(cocktailCommentDTO.toString());
         try {
-            boolean res = cocktailService.createComment(cocktailCommentDTO);
-            return ResponseEntity.ok().body(Map.of("success", "true"));
+            Long uid = Long.valueOf((String) request.getAttribute("uid"));
+            if(uid != null) {
+                boolean res = cocktailService.createComment(cocktailCommentDTO);
+                return ResponseEntity.ok().body(Map.of("success", res));
+            } else {
+                return ResponseEntity.badRequest().body(Map.of("success", "Authorization is null"));
+            }
         } catch (Exception e) {
             e.printStackTrace();
             return ResponseEntity.badRequest().body(Map.of("success", "false"));
@@ -141,23 +145,33 @@ public class CocktailController {
         }
     }
 
-    @PutMapping(value = "/comment")
-    public ResponseEntity updateReview(@RequestBody CocktailCommentDTO cocktailCommentDTO) {
+    @PutMapping(value = "/auth/comment")
+    public ResponseEntity updateReview(HttpServletRequest request, @RequestBody CocktailCommentDTO cocktailCommentDTO) {
         // commentId, comment 받음
         try {
-            cocktailService.updateComment(cocktailCommentDTO);
-            return ResponseEntity.ok().body(Map.of("success", "true"));
+            Long uid = Long.valueOf((String) request.getAttribute("uid"));
+            if(uid != null){
+                cocktailService.updateComment(cocktailCommentDTO);
+                return ResponseEntity.ok().body(Map.of("success", "true"));
+            } else {
+                return ResponseEntity.badRequest().body(Map.of("success", "Authorization is null"));
+            }
         } catch (Exception e) {
             e.printStackTrace();
             return ResponseEntity.badRequest().body(Map.of("success", "false"));
         }
     }
 
-    @DeleteMapping(value = "/comment")
-    public ResponseEntity deleteReview(@RequestBody CocktailCommentDTO cDTO) {
+    @DeleteMapping(value = "/auth/comment")
+    public ResponseEntity deleteReview(HttpServletRequest request, @RequestBody CocktailCommentDTO cDTO) {
         try {
-            boolean res = cocktailService.deleteComment(cDTO.getUserId(), cDTO.getCommentId());
-            return ResponseEntity.ok().body(Map.of("success", res));
+            Long uid = Long.valueOf((String) request.getAttribute("uid"));
+            if(uid != null){
+                boolean res = cocktailService.deleteComment(cDTO.getUserId(), cDTO.getCommentId());
+                return ResponseEntity.ok().body(Map.of("success", res));
+            } else {
+            return ResponseEntity.badRequest().body(Map.of("success", "Authorization is null"));
+            }
         } catch (Exception e) {
             e.printStackTrace();
             return ResponseEntity.badRequest().body(Map.of("success", "false"));
